refactored Taxon, Assembly, GenomeAnnotation API classes into Abstract, Proxy and subclasses
