+++ lib/biokbase/data_api/tests/shared.py
@@ -6,23 +6,23 @@ __date__ = '8/11/15'
 
 import logging
 
-from biokbase.data_api.assembly import AssemblyAPI
+from biokbase.data_api.core import ObjectAPI
 
-can_connect = False
 services = {
     "workspace_service_url": "https://ci.kbase.us/services/ws/",
     "shock_service_url": "https://ci.kbase.us/services/shock-api/",
 }
 genome = 'PrototypeReferenceGenomes/kb|g.3157'
 
-def setup():
-    global can_connect
+def can_connect():
+    try:
+        _ = ObjectAPI(services=services, ref=genome)
+        return True
+    except Exception, e:
+        _log.exception(e)
+        _log.warn('Cannot connect to workspace! Most tests will be skipped')
+        return False
 
+def setup():
     logging.basicConfig()
     _log = logging.getLogger()
-
-    try:
-        _ = AssemblyAPI(services=services, ref=genome + '_assembly')
-        can_connect = True
-    except:
-        _log.warn('Cannot connect to workspace! Most tests will be skipped')
