+++ setup.py
@@ -25,6 +25,9 @@ config = {
     "install_requires": parse_requirements(),
     "packages": ["biokbase",
                  "biokbase.data_api",
+                 "biokbase.data_api.annotation",
+                 "biokbase.data_api.sequence",
+                 "biokbase.data_api.taxonomy",
                  "biokbase.data_api.tests",
                  "biokbase.data_api.tests.performance",
                  "biokbase.workspace"],
